Move list_screen.dart to screens subfolder
