+++ lib/list/delete_list.dart
@@ -1,8 +1,9 @@
-import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
+
 import 'package:provider/provider.dart';
+
 import 'package:shopping_list/firestore/firestore_user.dart';
-import 'package:shopping_list/list/list_screen.dart';
+import 'package:shopping_list/list/screens/list_screen.dart';
 
 /// Deletes a list document in Firebase.
 deleteList({@required BuildContext context, @required String listID}) {
