+++ lib/loading_screen.dart
@@ -7,7 +7,7 @@ import 'package:shopping_list/authentication/authentication.dart';
 import 'package:shopping_list/authentication/screens/signin_screen.dart';
 import 'package:shopping_list/firestore/firestore_user.dart';
 import 'package:shopping_list/globals.dart';
-import 'package:shopping_list/list/list_screen.dart';
+import 'package:shopping_list/list/screens/list_screen.dart';
 
 class LoadingScreen extends StatefulWidget {
   @override
