+++ lib/main.dart
@@ -8,7 +8,7 @@ import 'package:shopping_list/authentication/screens/email_signin_screen.dart';
 import 'package:shopping_list/authentication/screens/signin_screen.dart';
 import 'package:shopping_list/firestore/firestore_user.dart';
 import 'package:shopping_list/globals.dart';
-import 'package:shopping_list/list/list_screen.dart';
+import 'package:shopping_list/list/screens/list_screen.dart';
 import 'package:shopping_list/loading_screen.dart';
 
 void main() async {
